Exception handling - delete

+++ src/main/java/free/biava/course/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import free.biava.course.entities.User;
 import free.biava.course.repositories.UserRepository;
+import free.biava.course.services.exceptions.DatabaseException;
 import free.biava.course.services.exceptions.ResourceNotFoundException;
 
 @Service
@@ -30,7 +33,18 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		repository.deleteById(id);
+		if(repository.existsById(id)) {
+			try {
+				repository.deleteById(id);
+			}
+			catch(DataIntegrityViolationException e) {
+				throw new DatabaseException(e.getMessage());
+			}
+		}
+		else {
+			throw new ResourceNotFoundException(id);
+		}
+		
 	}
 	
 	public User update(Long id, User obj) {
